refactor: rename files to follow convention
