+++ libs/methodologies/bold/rule-processors/mass-id/weighing/src/scale-ticket-verification/scale-ticket-verification.helpers.ts
