+++ libs/shared/scale-ticket-extractor/src/layout-1/index.ts
@@ -1,2 +1,2 @@
-export * from './parser';
-export * from './types';
+export * from './scale-ticket-layout-1.parser';
+export * from './scale-ticket-layout-1.types';
