+++ libs/shared/scale-ticket-extractor/src/layout-1/scale-ticket-layout-1.helpers.ts
