+++ libs/shared/scale-ticket-extractor/src/layout-1/scale-ticket-layout-1.types.ts
@@ -1,6 +1,6 @@
 import type { NonEmptyString } from '@carrot-fndn/shared/types';
 
-import type { ScaleTicketData } from '../types';
+import type { ScaleTicketData } from '../scale-ticket.types';
 
 export interface Layout1ScaleTicketData extends ScaleTicketData {
   transporter:
