+++ libs/shared/scale-ticket-extractor/src/scale-ticket.types.ts
