+++ libs/shared/text-extractor/src/index.ts
@@ -1,2 +1,2 @@
 export { textExtractor } from './text-extractor.provider';
-export * from './types';
+export * from './text-extractor.types';
