+++ libs/shared/text-extractor/src/text-extractor.provider.ts
@@ -1,6 +1,6 @@
 import { TextractClient } from '@aws-sdk/client-textract';
 
-import type { TextExtractor } from './types';
+import type { TextExtractor } from './text-extractor.types';
 
 import { TextractService } from './textract.service';
