+++ libs/shared/text-extractor/src/text-extractor.types.ts
